Move selector functions to selectors.js
